Create autoMusic.py
I'm trying to learn how to use Python to automate certain tasks to increase my workflow efficiency. This module is meant to play music in the background by interacting with the website Band Camp on its own. I would like to eventually add the ability to do the same with other websites, such as Youtube.